Refactor field naming rules to use a common base, and exclude Const fields
