+++ src/Tests/SonarLint.UnitTest/TestCases/PublicFieldName.vb
@@ -6,5 +6,6 @@
     Public Shared FooFooFFFoooFF As Integer
     Friend ReadOnly FooFooFFFFoooFF As Integer ' Noncompliant
     Protected FooFooFFFFoooFFF As Integer ' Noncompliant
+    Protected Const FooFooFFFFoooFFFF As Integer
 
 End Class
